+++ paddlenlp/ops/faster_transformer/src/pd_traits.h
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License. */
 #pragma once
 
-#include "fastertransformer/common.h"
+#include "fastertransformer/utils/common.h"
 
 using namespace fastertransformer;
